makepkg: run locally with libtool style wrapper
Build makepkg to scripts/.lib/makepkg and add a wrapper script to
call it. This is not useful at the moment, but is the first step
to allowing makepkg to be split into smaller pieces.

Signed-off-by: Allan McRae <allan@archlinux.org>
(cherry picked from commit 015667ab2c8f0be3b5ff34cc93757e7873a09dbd)
Signed-off-by: Ashley Whetter <ashley@awhetter.co.uk>

Conflicts:
	scripts/.gitignore
	scripts/Makefile.am